Ajustando tipo do campo Quantidade

+++ ControleEstoque.Api/Model/Response/TotalEstoqueResponse.cs
@@ -7,7 +7,7 @@ namespace ControleEstoque.Api.Model.Response
     public class TotalEstoqueResponse
     {
         [JsonProperty("quantidade")]
-        public double Quantidade { get; set; }
+        public int Quantidade { get; set; }
 
         [JsonProperty("total")]
         public double Total { get; set; }
